Compiler documentation, fixes, namespace

+++ Compiler/Src/Utility.h
@@ -40,11 +40,11 @@ using namespace std::string_literals;
 using namespace std::chrono_literals;
 
 using u8  = uint8_t ;
-using u16 = uint32_t;
+using u16 = uint16_t;
 using u32 = uint32_t;
 using u64 = uint64_t;
 using i8  =  int8_t ;
-using i16 =  int32_t;
+using i16 =  int16_t;
 using i32 =  int32_t; 
 using i64 =  int64_t;
 using f32 =  float  ;
